pyproject.toml
[tool.poetry]
name = "greenwave"
version = "2.2.0"
description = "Decision-making service for gating in a software delivery pipeline"
authors = [
  "Red Hat, Inc. and others"
]
license = "GPL-2.0-or-later"
readme = "README.md"
repository = "https://github.com/release-engineering/greenwave"
homepage = "https://gating-greenwave.readthedocs.io/"

include = [
    "COPYING",
    "README.md",
    "pytest.ini",
    "conftest.py",
    "run-dev-server.py",
    "tox.ini",
    "conf/*",
    "docs/*",
]

[tool.poetry.dependencies]
python = ">=3.8,<3.13"
PyYAML = "^6.0"
"dogpile.cache" = "^1.3.0"
flask = "^3.0.0"
gunicorn = "^21.2.0"
requests = "^2.31.0"
python-dateutil = "^2.8.2"
fedora-messaging = "^3.4.1"

# latest versions of stomp.py 8.0.0 and 8.0.1 are causing connection issues
"stomp.py" = "^8.1.0"

flake8 = {version = "^5.0.4", optional = true}
pytest = {version = "^7.4.4", optional = true}
pytest-cov = {version = "^4.1.0", optional = true}
mock = {version = "^5.1.0", optional = true}
requests-mock = {version = "^1.11.0", optional = true}

SQLAlchemy = {version = "^2.0.24", optional = true}
psycopg2-binary = {version = "^2.9.9", optional = true}

sphinx = {version = "^7.1.2", optional = true}
sphinxcontrib-httpdomain = {version = "^1.8.1", optional = true}

statsd = "^4.0.1"
pymemcache = "^4.0.0"
defusedxml = "^0.7.1"

opentelemetry-sdk = "^1.20.0"
opentelemetry-instrumentation-flask = "^0.41b0"
opentelemetry-instrumentation-requests = "^0.41b0"
opentelemetry-exporter-otlp-proto-http = "^1.20.0"

[tool.poetry.extras]
test = [
    "flake8",
    "pytest",
    "pytest-cov",
    "mock",
]
functional-test = [
    "SQLAlchemy",
    "psycopg2-binary",
]
docs = [
    "sphinx",
    "sphinxcontrib-httpdomain",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
